+++ packages/jaspr/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased minor
 
+- Allow setting `Document(base: )` to `null`, and fix the path to the generated client script when no `<base>` element exists.
+- Added `--include-source-maps` option to `jaspr build` command.
 - Fixed sitemap generation bug in top-level route ('/')
 - Update `build` dependency to 3.0.0 and migrate to new analyzer model.
 - Added `jaspr tooling-daemon` command and component scope analysis.